added understand score, and refresh button on teachers view

+++ migrations/20131009124559-transcripts.js
@@ -6,7 +6,7 @@ exports.up = function (db, callback) {
     id: { type: 'int', primaryKey: true, autoIncrement: true },
     transcript: 'string',
     session: 'int',
-    time: 'int'
+    time: 'bigint'
   }, callback);
 };
